requirements.txt
boto3==1.28.29
boto3-stubs==1.28.29
botocore==1.31.29
botocore-stubs==1.31.29
jmespath==1.0.1
mypy-boto3-cloudformation==1.28.19
mypy-boto3-dynamodb==1.28.27
mypy-boto3-ec2==1.28.29
mypy-boto3-lambda==1.28.19
mypy-boto3-rds==1.28.19
mypy-boto3-s3==1.28.27
mypy-boto3-sqs==1.28.19
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
types-awscrt==0.19.0
types-s3transfer==0.6.2
urllib3==1.26.16
pyyaml==6.0.1
jsonschema==4.19.0